// src/app/Models/daramamaster.ts
export class DramaMAsterr {

    ID: number = 0;
    SHORT_CODE: any
    NAME: string = '';
    CAST_NAMES: any = []
    DESCRIPTION: string = '';
    DRAMA_IMAGE: string = '';
    STATUS: boolean = true;
    SEQUENCE_NUMBER: number = 0;
    FROM_DATE: any;
    TO_DATE: any;
    CITY_ID: number = 0;
    CITY_NAME: string = '';

}